+++ packages/core/nuxt-theme-module/theme/components/AppHeader.vue
@@ -20,60 +20,7 @@
         </nuxt-link>
       </template>
       <template #navigation>
-        <SfHeaderNavigation>
-          <SfHeaderNavigationItem
-            v-for="(category, index) in categories"
-            :key="index"
-            :label="category.name"
-            @mouseenter="hovered = category.slug"
-            @mouseleave="hovered = ''"
-            @click="hovered = ''"
-            :link="localePath(`/c/${category.slug}`)"
-          >
-            <SfMegaMenu
-              :is-absolute="true"
-              :visible="hovered === category.slug"
-              :title="category.name"
-              @close="hovered = ''"
-              v-if="category && category.children.length"
-            >
-              <SfMegaMenuColumn
-                v-for="(subCategory, subIndex) in category.children"
-                :key="subIndex"
-                :title="subCategory.name"
-              >
-                <SfList>
-                  <SfListItem
-                    v-for="(subCategoryChild, childIndex) in subCategory.children"
-                    :key="childIndex"
-                  >
-                    <SfMenuItem :label="subCategoryChild.name" :link="localePath(`/c/${subCategoryChild.slug}`)">
-                      <SfLink>
-                        {{ subCategoryChild.name }}
-                      </SfLink>
-                    </SfMenuItem>
-                  </SfListItem>
-                </SfList>
-              </SfMegaMenuColumn>
-              <SfMegaMenuColumn
-                v-if="isCategoryWithBanners"
-                title="Featured"
-                class="sf-mega-menu-column--pined-content-on-mobile sf-mega-menu-column--hide-header-on-mobile sb-mega-menu__featured"
-              >
-                <div class="sb-mega-menu__banners">
-                  <SfBanner
-                    v-for="(banner, key) in banners"
-                    :key="key"
-                    :title="banner.title"
-                    :subtitle="banner.subtitle"
-                    :image="banner.pictures"
-                    class="sb-mega-menu__banner"
-                  />
-                </div>
-              </SfMegaMenuColumn>
-            </SfMegaMenu>
-          </SfHeaderNavigationItem>
-        </SfHeaderNavigation>
+        <TopMenu @setOverlay="hovered = $event" />
       </template>
       <template #aside>
         <LocaleSelector class="smartphone-only" />
@@ -83,25 +30,23 @@
 </template>
 
 <script>
-import { SfHeader, SfImage, SfMegaMenu, SfList, SfLink, SfMenuItem, SfBanner, SfOverlay } from '@storefront-ui/vue';
+import { SfHeader, SfImage, SfLink, SfOverlay } from '@storefront-ui/vue';
 import { useUiState } from '~/composables';
-import { useCart, useWishlist, useUser, useCategory, cartGetters } from '<%= options.generate.replace.composables %>';
+import { useCart, useWishlist, useUser, cartGetters } from '<%= options.generate.replace.composables %>';
 import { computed, ref } from '@vue/composition-api';
 import { onSSR } from '@vue-storefront/core';
 import { useUiHelpers } from '~/composables';
 import LocaleSelector from './LocaleSelector';
+import TopMenu from './TopMenu';
 
 export default {
   components: {
     SfHeader,
     SfImage,
-    SfMegaMenu,
-    SfList,
     SfLink,
-    SfBanner,
-    SfMenuItem,
     SfOverlay,
-    LocaleSelector
+    LocaleSelector,
+    TopMenu
   },
   setup(props, { root }) {
     const { toggleCartSidebar, toggleWishlistSidebar, toggleLoginModal } = useUiState();
@@ -109,10 +54,8 @@ export default {
     const { isAuthenticated, load } = useUser();
     const { cart, loadCart } = useCart();
     const { loadWishlist } = useWishlist();
-    const { categories, search } = useCategory('menu-categories');
     const term = ref(getFacetsFromURL().term);
-    const hovered = ref('');
-    const categoriesWithBanners = ref(['new']);
+    const hovered = ref(false);
 
     const cartTotalItems = computed(() => {
       const count = cartGetters.getTotalItems(cart.value);
@@ -120,8 +63,6 @@ export default {
     });
 
     const accountIcon = computed(() => isAuthenticated.value ? 'profile_fill' : 'profile');
-
-    const isCategoryWithBanners = computed(() => categoriesWithBanners.value.includes(hovered.value));
 
     // TODO: https://github.com/DivanteLtd/vue-storefront/issues/4927
     const handleAccountClick = async () => {
@@ -136,7 +77,6 @@ export default {
       await load();
       await loadCart();
       await loadWishlist();
-      await search({});
     });
 
     return {
@@ -147,32 +87,7 @@ export default {
       toggleWishlistSidebar,
       changeSearchTerm,
       term,
-      hovered,
-      categoriesWithBanners,
-      isCategoryWithBanners,
-      categories
-    };
-  },
-  data() {
-    return {
-      banners: [
-        {
-          title: 'THE OFFICE LIFE',
-          subtitle: 'T-shirts',
-          pictures: {
-            mobile: '/megamenu/bannerA.webp',
-            desktop: '/megamenu/bannerA.webp'
-          }
-        },
-        {
-          title: 'ECO SANDALS',
-          subtitle: 'T-shirts',
-          pictures: {
-            mobile: '/megamenu/bannerB.webp',
-            desktop: '/megamenu/bannerB.webp'
-          }
-        }
-      ]
+      hovered
     };
   }
 };
@@ -192,28 +107,5 @@ export default {
 
 .nav-item {
   --header-navigation-item-margin: 0 var(--spacer-base);
-}
-
-.sb-mega-menu {
-  &__featured {
-    flex: 0 0 43.125rem;
-  }
-  &__banners {
-    display: flex;
-    flex-direction: column;
-    padding: var(--spacer-base);
-    @include for-desktop {
-      flex-direction: row;
-      padding: 0;
-    }
-  }
-  &__banner{
-    &:first-child{
-      margin: 0 0 var(--spacer-sm) 0;
-      @include for-desktop {
-        margin: 0 var(--spacer-sm) 0 0;
-      }
-    }
-  }
 }
 </style>
